tetris: adds basic bag support
with default settings (without `-b` flag) a bag containing each tile
exactly once is generated. The order inside this bag is random, tiles
are dealt in order. It is now impossible to get the same tile three
times.

The old random generator is still available with the `-b` flag. With the
new code it is no longer a #define, but part of the function.